v2ray

+++ v2ray.sh
@@ -28,11 +28,22 @@ if [ ! -f "$file" ]; then
 
     /bin/systemctl enable v2ray
 
+    echo "v2ray install success"
+
+fi
+
+bbr=`lsmod|grep bbr|awk '{print $1}'`
+
+if [[ "tcp_bbr" != $bbr ]]; then
+
+    bbr = `lsmod|grep bbr|awk '{print $1}'`
+
+    curl -L -s https://raw.githubusercontent.com/qingfenghuohu/os_init/master/bbr.sh | bash
+
+    reboot
+
+    echo "bbr install success"
+
 fi
 
 
-curl -L -s https://raw.githubusercontent.com/qingfenghuohu/os_init/master/bbr.sh | bash
-
-reboot
-
-echo "success"
